Played around with single player games. We now solve kitten extremely fast, aipsrovers by simulating and getting a solution there. In hanoi we at least take away some points. Yippeeh.

+++ src/player/IHeuristic.java
@@ -4,5 +4,5 @@ package player;
 import org.eclipse.palamedes.gdl.core.model.IGameNode;
 
 public interface IHeuristic {
-	public double calculateHeuristic(IGameNode node);
+	public int calculateHeuristic(IGameNode node);
 }
