+++ src/player/InverseMobilityHeuristic.java
@@ -8,7 +8,7 @@ public class InverseMobilityHeuristic implements IHeuristic {
 	private int roleIndex;
 	
 	@Override
-	public double calculateHeuristic(IGameNode node) {
+	public int calculateHeuristic(IGameNode node) {
 		int myMoves = node.getState().getLegalMoves()[roleIndex].length;
 		int movesOfTheOthers = 0;
 		for(int i=0; i<node.getState().getLegalMoves().length; i++) {
